# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HUTCH Resolve planned API",
    "version": "1.0.0",
    "description": "Shared Resolve wire contract v1.0.0. Implemented routes include health/readiness, session lifecycle, customer account/case/investigation APIs, action/receipt, agent review, and signed Voice session/callback endpoints. Conversation turns depend on Tevin's injected in-process controller; full frontends and live Voice qualification remain pending. See docs/contracts.md and context.md for status."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Resolve runtime"
    }
  ],
  "paths": {
...
    "/cases/{id}/action-proposals": {
      "post": {
        "operationId": "propose_action",
        "summary": "propose action",
        "description": "Proposed interface; not implemented at documentation baseline.",
        "tags": [
          "Resolve"
        ],
        "x-owner": "Harry",
        "security": [
          {
            "CustomerSession": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Correlation ID"
              }
            }
          },
          "401": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          }
        }
      }
    },
    "/action-proposals/{id}/confirmations": {
      "post": {
        "operationId": "confirm_or_decline",
        "summary": "confirm or decline",
        "description": "Customer confirmation is idempotent. ACCEPT is durably queued only when action execution is available; otherwise returns retryable 503 ACTION_EXECUTION_UNAVAILABLE without recording the acceptance. DECLINE remains available.",
        "tags": [
          "Resolve"
        ],
        "x-owner": "Harry",
        "security": [
          {
            "CustomerSession": []
          }
        ],
        "responses": {
          "202": {
            "description": "Durably accepted; not completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationResult"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Correlation ID"
              }
            }
          },
          "401": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Action execution unavailable or dependency unavailable; see Error.code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Decline persisted; operation is null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationResult"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationRequest"
              }
            }
          }
        }
      }
    },
    "/operations/{id}": {
      "get": {
        "operationId": "get_operation",
        "summary": "get operation",
        "description": "Proposed interface; not implemented at documentation baseline.",
        "tags": [
          "Resolve"
        ],
        "x-owner": "Harry",
        "security": [
          {
            "CustomerSession": []
          },
          {
            "AgentSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationView"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Correlation ID"
              }
            }
          },
          "401": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/ResolveRealm"
          }
        ]
      }
    },
    "/cases/{id}/escalations": {
      "post": {
        "operationId": "prepare_escalation",
        "summary": "prepare escalation",
        "description": "Persist the reason with a CREATE_REVIEW_TICKET proposal; normal customer confirmation is still required before delivery.",
        "tags": [
          "Resolve"
        ],
        "x-owner": "Harry",
        "security": [
          {
            "CustomerSession": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Correlation ID"
              }
            }
          },
          "401": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalationRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CustomerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "resolve_customer_session"
      },
      "AgentSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "resolve_agent_session"
      },
      "VoiceSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Voice-Signature",
        "description": "HMAC SHA256 of timestamp.event_id.body_sha256; all companion headers mandatory; 60-second skew."
      }
    },
    "schemas": {
...
      "ActionType": {
        "type": "string",
        "enum": [
          "DEACTIVATE_VAS",
          "SEND_SETTINGS_INSTRUCTIONS",
          "CREATE_REVIEW_TICKET"
        ]
      },
      "OperationStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "SUCCEEDED",
          "FAILED",
          "UNKNOWN",
          "REVIEW_REQUIRED"
        ]
      },
...
      "Decision": {
        "type": "string",
        "enum": [
          "ACCEPT",
          "DECLINE"
        ]
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "request_id": {
                "type": "string",
                "format": "uuid"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "retryable",
              "request_id",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "ProposalRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "investigation_id": {
            "type": "string",
            "format": "uuid"
          },
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "target_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "expected_version",
          "investigation_id",
          "action_type",
          "target_id"
        ]
      },
      "ProposalView": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "investigation_id": {
            "type": "string",
            "format": "uuid"
          },
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "target_id": {
            "type": "string",
            "format": "uuid"
          },
          "target_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "target_label": {
            "type": "string"
          },
          "consequences": {
            "type": "string"
          },
          "proposal_hash": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "simulation": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "id",
          "case_id",
          "investigation_id",
          "action_type",
          "target_id",
          "target_version",
          "target_label",
          "consequences",
          "proposal_hash",
          "expires_at",
          "simulation"
        ]
      },
      "ConfirmationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "proposal_hash": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "client_turn_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "proposal_hash",
          "decision",
          "client_turn_id"
        ]
      },
      "ConfirmationView": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "proposal_id": {
            "type": "string",
            "format": "uuid"
          },
          "proposal_hash": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "channel": {
            "type": "string",
            "enum": [
              "TEXT",
              "VOICE",
              "AGENT"
            ]
          },
          "client_turn_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "operation_status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PENDING"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "simulation": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "id",
          "proposal_id",
          "proposal_hash",
          "decision",
          "channel",
          "client_turn_id",
          "created_at",
          "operation_id",
          "operation_status",
          "simulation"
        ]
      },
      "OperationOutcome": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "actual_target_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider_ticket_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "code",
          "message",
          "actual_target_status",
          "provider_ticket_id"
        ]
      },
      "OperationView": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "proposal_id": {
            "type": "string",
            "format": "uuid"
          },
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "status": {
            "$ref": "#/components/schemas/OperationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "provider_operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "outcome": {
            "$ref": "#/components/schemas/OperationOutcome"
          },
          "next_step": {
            "type": "string"
          },
          "simulation": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "id",
          "case_id",
          "proposal_id",
          "action_type",
          "status",
          "created_at",
          "updated_at",
          "provider_operation_id",
          "outcome",
          "next_step",
          "simulation"
        ]
      },
      "ConfirmationResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "confirmation": {
            "$ref": "#/components/schemas/ConfirmationView"
          },
          "operation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperationView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "confirmation",
          "operation"
        ]
      },
      "EscalationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "investigation_id": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "expected_version",
          "investigation_id",
          "reason"
        ]
      },
...
    },
    "parameters": {
      "ResolveRealm": {
        "name": "X-Resolve-Realm",
        "in": "header",
        "required": false,
        "description": "Selects the browser session realm for shared customer/agent reads. Required when both cookies are present.",
        "schema": {
          "type": "string",
          "enum": [
            "customer",
            "agent"
          ]
        }
      }
    }
  }
}